CUDA subdir cleanup

+++ cuda/cuda/fem/bilinearform.cpp
@@ -1,13 +1,19 @@
-// Copyright (c) 2010, Lawrence Livermore National Security, LLC. Produced at
-// the Lawrence Livermore National Laboratory. LLNL-CODE-443211. All Rights
-// reserved. See file COPYRIGHT for details.
+// Copyright (c) 2017, Lawrence Livermore National Security, LLC. Produced at
+// the Lawrence Livermore National Laboratory. LLNL-CODE-734707. All Rights
+// reserved. See files LICENSE and NOTICE for details.
 //
-// This file is part of the MFEM library. For more information and source code
-// availability see http://mfem.org.
+// This file is part of CEED, a collection of benchmarks, miniapps, software
+// libraries and APIs for efficient high-order finite element and spectral
+// element discretizations for exascale applications. For more information and
+// source code availability see http://github.com/ceed.
 //
-// MFEM is free software; you can redistribute it and/or modify it under the
-// terms of the GNU Lesser General Public License (as published by the Free
-// Software Foundation) version 2.1 dated February 1999.
+// The CEED research is supported by the Exascale Computing Project 17-SC-20-SC,
+// a collaborative effort of two U.S. Department of Energy organizations (Office
+// of Science and the National Nuclear Security Administration) responsible for
+// the planning and preparation of a capable exascale ecosystem, including
+// software, applications, hardware, advanced system engineering and early
+// testbed platforms, in support of the nation's exascale computing imperative.
+
 #include "../cuda.hpp"
 
 namespace mfem
@@ -64,8 +70,8 @@ void CudaBilinearForm::AddIntegrator(CudaIntegrator* i,
 // ***************************************************************************
 void CudaBilinearForm::Assemble()
 {
-   const int integratorCount = (int) integrators.size();
-   for (int i = 0; i < integratorCount; ++i)
+   const size_t integratorCount = integrators.size();
+   for (size_t i = 0; i < integratorCount; ++i)
    {
       integrators[i]->Assemble();
    }
@@ -132,8 +138,8 @@ void CudaBilinearForm::Mult(const CudaVector& x, CudaVector& y) const
 {
    trialFes->GlobalToLocal(x, localX);
    localY = 0;
-   const int integratorCount = (int) integrators.size();
-   for (int i = 0; i < integratorCount; ++i)
+   const size_t integratorCount = integrators.size();
+   for (size_t i = 0; i < integratorCount; ++i)
    {
       integrators[i]->MultAdd(localX, localY);
    }
@@ -145,8 +151,8 @@ void CudaBilinearForm::MultTranspose(const CudaVector& x, CudaVector& y) const
 {
    testFes->GlobalToLocal(x, localX);
    localY = 0;
-   const int integratorCount = (int) integrators.size();
-   for (int i = 0; i < integratorCount; ++i)
+   const size_t integratorCount = integrators.size();
+   for (size_t i = 0; i < integratorCount; ++i)
    {
       integrators[i]->MultTransposeAdd(localX, localY);
    }
