+++ cuda/cuda/general/array.hpp
@@ -16,6 +16,12 @@
 #ifndef LAGHOS_CUDA_ARRAY
 #define LAGHOS_CUDA_ARRAY
 
+#include <cstddef>
+#include <cassert>
+#include "general/array.hpp"
+#include "../general/malloc.hpp"
+#include "../general/memcpy.hpp"
+
 namespace mfem
 {
 
@@ -26,11 +32,11 @@ template <class T> class CudaArray<T,true> : public rmalloc<T>
 {
 private:
    T* data = NULL;
-   size_t sz,d[4];
+   int sz,d[4];
 public:
    CudaArray():data(NULL),sz(0),d{0,0,0,0} {}
-   CudaArray(const size_t x) {allocate(x);}
-   CudaArray(const size_t x,const size_t y) {allocate(x,y);}
+   CudaArray(const int x) {allocate(x);}
+   CudaArray(const int x,const int y) {allocate(x,y);}
    CudaArray(const CudaArray<T,true> &r) {assert(false);}
    CudaArray& operator=(Array<T> &a)
    {
@@ -38,30 +44,30 @@ public:
       return *this;
    }
    ~CudaArray() {rmalloc<T>::operator delete (data);}
-   inline size_t* dim() { return &d[0]; }
+   inline int* dim() { return &d[0]; }
    inline T* ptr() { return data; }
    inline const T* GetData() const { return data; }
    inline const T* ptr() const { return data; }
    inline operator T* () { return data; }
    inline operator const T* () const { return data; }
    double operator* (const CudaArray& a) const { return vector_dot(sz, data, a.data); }
-   inline size_t size() const { return sz; }
-   inline size_t Size() const { return sz; }
-   inline size_t bytes() const { return size()*sizeof(T); }
-   void allocate(const size_t X, const size_t Y =1,
-                 const size_t Z =1, const size_t D =1,
+   inline int size() const { return sz; }
+   inline int Size() const { return sz; }
+   inline int bytes() const { return size()*sizeof(T); }
+   void allocate(const int X, const int Y =1,
+                 const int Z =1, const int D =1,
                  const bool transposed = false)
    {
       d[0]=X; d[1]=Y; d[2]=Z; d[3]=D;
       sz=d[0]*d[1]*d[2]*d[3];
       data=(T*) rmalloc<T>::operator new (sz);
    }
-   inline T& operator[](const size_t x) { return data[x]; }
-   inline T& operator()(const size_t x, const size_t y)
+   inline T& operator[](const int x) { return data[x]; }
+   inline T& operator()(const int x, const int y)
    {
       return data[x + d[0]*y];
    }
-   inline T& operator()(const size_t x, const size_t y, const size_t z)
+   inline T& operator()(const int x, const int y, const int z)
    {
       return data[x + d[0]*(y + d[1]*z)];
    }
@@ -69,7 +75,7 @@ public:
    {
       T *h_data = (double*) ::malloc(bytes());
       rmemcpy::rDtoH(h_data,data,bytes());
-      for (size_t i=0; i<sz; i+=1)
+      for (int i=0; i<sz; i+=1)
          if (sizeof(T)==8) { printf("\n\t[%ld] %.15e",i,h_data[i]); }
          else { printf("\n\t[%ld] %d",i,h_data[i]); }
       free(h_data);
@@ -82,10 +88,10 @@ template <class T> class CudaArray<T,false> : public rmalloc<T>
 private:
    static const int DIM = 4;
    T* data = NULL;
-   size_t sz,d[DIM];
+   int sz,d[DIM];
 public:
    CudaArray():data(NULL),sz(0),d{0,0,0,0} {}
-   CudaArray(const size_t d0) {allocate(d0);}
+   CudaArray(const int d0) {allocate(d0);}
    CudaArray(const CudaArray<T,false> &r) {assert(false);}
    ~CudaArray() {rmalloc<T>::operator delete (data);}
    CudaArray& operator=(Array<T> &a)
@@ -93,18 +99,18 @@ public:
       rmemcpy::rHtoD(data,a.GetData(),a.Size()*sizeof(T));
       return *this;
    }
-   inline size_t* dim() { return &d[0]; }
+   inline int* dim() { return &d[0]; }
    inline T* ptr() { return data; }
    inline T* GetData() const { return data; }
    inline const T* ptr() const { return data; }
    inline operator T* () { return data; }
    inline operator const T* () const { return data; }
    double operator* (const CudaArray& a) const { return vector_dot(sz, data, a.data); }
-   inline size_t size() const { return sz; }
-   inline size_t Size() const { return sz; }
-   inline size_t bytes() const { return size()*sizeof(T); }
-   void allocate(const size_t X, const size_t Y =1,
-                 const size_t Z =1, const size_t D =1,
+   inline int size() const { return sz; }
+   inline int Size() const { return sz; }
+   inline int bytes() const { return size()*sizeof(T); }
+   void allocate(const int X, const int Y =1,
+                 const int Z =1, const int D =1,
                  const bool transposed = false)
    {
       d[0]=X; d[1]=Y; d[2]=Z; d[3]=D;
@@ -113,19 +119,19 @@ public:
       data=(T*) rmalloc<T>::operator new (sz);
 #define xsw(a,b) a^=b^=a^=b
       if (transposed) { xsw(d[0],d[1]); }
-      for (size_t i=1,b=d[0]; i<DIM; xsw(d[i],b),++i)
+      for (int i=1,b=d[0]; i<DIM; xsw(d[i],b),++i)
       {
          d[i]*=d[i-1];
       }
       d[0]=1;
       if (transposed) { xsw(d[0],d[1]); }
    }
-   inline T& operator[](const size_t x) { return data[x]; }
-   inline T& operator()(const size_t x, const size_t y)
+   inline T& operator[](const int x) { return data[x]; }
+   inline T& operator()(const int x, const int y)
    {
       return data[d[0]*x + d[1]*y];
    }
-   inline T& operator()(const size_t x, const size_t y, const size_t z)
+   inline T& operator()(const int x, const int y, const int z)
    {
       return data[d[0]*x + d[1]*y + d[2]*z];
    }
@@ -133,7 +139,7 @@ public:
    {
       T *h_data = (double*) ::malloc(bytes());
       rmemcpy::rDtoH(h_data,data,bytes());
-      for (size_t i=0; i<sz; i+=1)
+      for (int i=0; i<sz; i+=1)
          if (sizeof(T)==8) { printf("\n\t[%ld] %.15e",i,h_data[i]); }
          else { printf("\n\t[%ld] %d",i,h_data[i]); }
       free(h_data);
